updated custom check boxes tree
